requirements.txt
pywebview
pyautogui
pygetwindow
pywin32